refactor: move the bank transactions header and chrome into the domain
BankTransactionsHeader/ takes BankTransactionsHeaderMenu as a private nested part
-- only the header can render it -- and absorbs BankTransactionsActions as
BankTransactionsHeaderActions. That was a 13-line div with one consumer, and the
bare plural "Actions" read as domain-wide when it is header-local. One directory
goes away.

BankTransactionsBulkActions/ moves to the root with its three modals and utils
nested.

Three renames for names that said nothing about bank transactions:
  TransactionsSearch      -> BankTransactionsSearchField  ("Transactions" is
    ambiguous across journal, custom-account, and bank feeds; Field matches @ui)
  SelectedBankAccountsChip -> BankTransactionsAccountFilterChip  (qualified, and
    "filter" says why it exists)
  SyncingComponent        -> BankTransactionsSyncingStatus  ("Component" as a
    suffix carries no information; "Status" says what it renders)

Every class-generating constant keeps its value, so no published selector moves:
CLASS_NAME stays 'Layer__BankTransactionsActions' in the renamed header actions
and 'Layer__bank-transactions__selected-accounts-chip' in the renamed chip, and
the syncing status still emits Layer__syncing-component, which
bankTransactionsHeader.scss targets. These deliberate name/class mismatches are
the input list for a follow-up BEM sweep.
