+++ src/components/BankTransactions/BankTransactions.tsx
@@ -30,11 +30,11 @@ import { BankTransactionsStringOverridesProvider } from '@contexts/BankTransacti
 import { CategorizationRulesContext, CategorizationRulesProvider } from '@contexts/CategorizationRulesContext/CategorizationRulesContext'
 import { InAppLinkProvider, type LinkingMetadata } from '@contexts/InAppLinkContext'
 import { Container } from '@blocks/Layout/Container/Container'
-import { BankTransactionsHeader } from '@components/BankTransactions/BankTransactionsHeader'
 import { BankTransactionsList } from '@components/BankTransactionsList/BankTransactionsList'
 import { BankTransactionsMobileList } from '@components/BankTransactionsMobileList/BankTransactionsMobileList'
 import { BankTransactionsTable } from '@components/BankTransactionsTable/BankTransactionsTable'
 import { ErrorBoundary } from '@components/utility/ErrorBoundary/ErrorBoundary'
+import { BankTransactionsHeader } from '@features/bankTransactions/BankTransactionsHeader/BankTransactionsHeader'
 import { BankTransactionsTableContent, type MobileComponentType } from '@features/bankTransactions/constants'
 import { type BankTransactionsStringOverrides } from '@features/bankTransactions/types'
 import { ResponsiveCategorizationRulesView } from '@features/categorization/ResponsiveCategorizationRulesView/ResponsiveCategorizationRulesView'
