+++ src/components/features/bankTransactions/BankTransactionsSyncingStatus/bankTransactionsSyncingStatus.scss
